+++ src/EntityGraphQL/Compiler/ConstantVisitor.cs
@@ -3,6 +3,8 @@ using System.Linq.Expressions;
 using EntityGraphQL.Grammer;
 using System.Text.RegularExpressions;
 using System.Globalization;
+using EntityGraphQL.Schema;
+using System.Linq;
 
 namespace EntityGraphQL.Compiler
 {
@@ -10,9 +12,11 @@ namespace EntityGraphQL.Compiler
     internal class ConstantVisitor : EntityGraphQLBaseVisitor<ExpressionResult>
     {
         public static readonly Regex GuidRegex = new Regex(@"^[0-9A-F]{8}[-]?([0-9A-F]{4}[-]?){3}[0-9A-F]{12}$", RegexOptions.IgnoreCase);
+        private readonly ISchemaProvider schema;
 
-        public ConstantVisitor()
+        public ConstantVisitor(ISchemaProvider schema)
         {
+            this.schema = schema;
         }
 
         public override ExpressionResult VisitInt(EntityGraphQLParser.IntContext context)
@@ -44,8 +48,20 @@ namespace EntityGraphQL.Compiler
 
         public override ExpressionResult VisitNull(EntityGraphQLParser.NullContext context)
         {
-            // we may need to convert a string into a DateTime or Guid type
             var exp = (ExpressionResult)Expression.Constant(null);
+            return exp;
+        }
+
+        public override ExpressionResult VisitIdentity(EntityGraphQLParser.IdentityContext context)
+        {
+            // this should be an enum
+            var enumVal = context.GetText();
+            var enumField = schema.EnumTypes()
+                .Select(e => e.GetFields().FirstOrDefault(f => f.Name == enumVal))
+                .Where(f => f != null)
+                .FirstOrDefault();
+
+            var exp = (ExpressionResult)Expression.Constant(Enum.Parse(enumField.ReturnTypeClr, enumField.Name));
             return exp;
         }
     }