+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Removed support for older syntax of complex queries that is not GQL standard
 - Refactored `GraphQLVistor` and friends to make it easier to follow what is happening (I hope). See CONTRIBUTING.md for some notes
 - From my testing the compiling and expression building is 15-20% faster than before (still network and or the DB calls are the largest)
+- Allow `enum` values in the query schema (e.g. as an argument)
 
 # 0.50.1
 - Name all `ParameterExpression`s as EF 3.1 expects a name (can throw an error)